% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_database.R
\name{create_database}
\alias{create_database}
\title{Creating database}
\usage{
create_database(dbname,user,password,host)
}
\arguments{
\item{dbname}{name of database}

\item{user}{name of user}

\item{password}{password of database}

\item{host}{name of host}
}
\value{
invisible NULL
}
\description{
Function \code{create_database} creates database with three empty
tables: deputies, votings, votes.
}
\details{
// to do
}
\examples{
// to do
}
\author{
Piotr Smuda
}

