% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deputies_get_data.R
\name{deputies_get_data}
\alias{deputies_get_data}
\title{Getting data about deputies}
\usage{
deputies_get_data(type)
}
\arguments{
\item{type}{type of deputies which be add to table with deputies: active, inactive}
}
\value{
data frame with two columns: id_deputy, surname_name
}
\description{
Function \code{deputies_get_data} gets data about deputies.
}
\details{
Function \code{deputies_get_data} gets deputies' ids and personal data like
name and surname. Also there is a choice between types of deputies, because
on the page of Polish diet deputies are splitted into \emph{active} and \emph{inactive}.
}
\note{
All information is stored in PostgreSQL database.
}
\examples{
\dontrun{
deputies_get_data('active')
deputies_get_data('inactive')}
}
\author{
Piotr Smuda
}

