% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_all_votes.R
\name{get_all_votes}
\alias{get_all_votes}
\title{Retrieve all votes from a database}
\usage{
get_all_votes(dbname = 'sejmrp', user = 'reader', password = 'qux94874', host = 'services.mini.pw.edu.pl')
}
\arguments{
\item{dbname}{name of database}

\item{user}{name of user}

\item{password}{password of database}

\item{host}{name of host}
}
\value{
a data.frame with  NULL
}
\description{
Function \code{get_all_votes} reads all votes from a database.
Default parameters use privilages of 'reader'. It can only SELECT data from database.
Note that, due to data size (~150 MB) it may take few seconds / minutes to download all votes.
}
\examples{
\dontrun{
all_votes <- get_all_votes()
dim(all_votes)
# [1] 2741899       8
object.size(all_votes)
# 144287808 bytes}
}
\author{
Przemyslaw Biecek
}

