% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_deputies_mds.R
\name{get_deputies_mds}
\alias{get_deputies_mds}
\title{Converts Deputies Distance Matrix into a Multidimensional Scaling Plot}
\usage{
get_deputies_mds(distances, clubs = NULL, plot = TRUE)
}
\arguments{
\item{distances}{a distance matrix, preferably created with \code{get_distance_matrix}}

\item{clubs}{a matrix that maps \code{deputy_id} (names from \code{distances}) on clubs}

\item{plot}{if \code{TRUE} then the \code{ggplot} object will be returned instead of the dendrogram itself}
}
\value{
MDS coordinates or a ggplot
}
\description{
Function \code{get_deputies_mds} converts the distance matrix between deputies 
into a 2D representation (Multidimensional Scaling)
}
\examples{
\dontrun{
votes <- get_deputies_mds(terms_of_office = c(7,7))
v <- c(`Za` = 5, `Przeciw` = -5, `Wstrzymał się` = 2, `Nieobecny` = 0)/10
mat2 <- get_distance_matrix(votes[,c("surname_name", "id_voting", "vote")], weights = v)
df <- votes[,c("surname_name","club")]
df \%>\%
  group_by(surname_name, club) \%>\%
  summarise(n = n()) \%>\% 
  arrange(-n) \%>\%
  group_by(surname_name) \%>\%
  top_n(1) \%>\%
  as.data.frame() -> clubs
row.names(clubs) <- clubs[,1]

get_deputies_mds(mat2, clubs)
}

}
\author{
Przemyslaw Biecek
}
