% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_deputies_silhouette.R
\name{get_deputies_silhouette}
\alias{get_deputies_silhouette}
\title{Converts Deputies Distance Matrix into a Silhouette Plot}
\usage{
get_deputies_silhouette(distances, clubs = NULL, plot = TRUE,
  remove_missing_clubs = TRUE)
}
\arguments{
\item{distances}{a distance matrix, preferably created with \code{get_distance_matrix}}

\item{clubs}{a data.frame that maps \code{deputy_id} (names from \code{distances}) on clubs. 
It should contain at least one column - \code{club} and rows should have names corresponding to names in the \code{distances}.}

\item{plot}{if \code{TRUE} then the \code{ggplot} object will be returned instead of the dendrogram itself}

\item{remove_missing_clubs}{if TRUE then rows of \code{distances} that are not mapped in \code{clubs} will be removed}
}
\value{
silhouette object or a ggplot
}
\description{
Function \code{get_deputies_silhouette} converts the distance matrix between deputies 
and their clubs into a silhouette plot
}
\examples{
library("dplyr")
# votes <- get_filtered_votes(terms_of_office = c(7,7))
data(votes)
v <- c(`Za` = 5, `Przeciw` = -5, `Wstrzymał się` = 2, `Nieobecny` = 0)/10
mat2 <- get_distance_matrix(votes[,c("surname_name", "id_voting", "vote")], weights = v)
df <- votes[,c("surname_name","club")]
df \%>\%
  group_by(surname_name, club) \%>\%
  summarise(n = n()) \%>\% 
  arrange(-n) \%>\%
  group_by(surname_name) \%>\%
  top_n(1) \%>\%
  as.data.frame() -> clubs
row.names(clubs) <- clubs[,1]
clubs$club[clubs$club == "niez."] = "cross-bencher"

get_deputies_silhouette(mat2, clubs)

}
\author{
Przemyslaw Biecek
}
