% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_distance_matrix.R
\name{get_distance_matrix}
\alias{get_distance_matrix}
\title{Converts Voting Table into the Deputies Distance Matrix}
\usage{
get_distance_matrix(votes, weights = NULL, allowMissings = 0)
}
\arguments{
\item{votes}{a data frame with three columns, respectively: deputy id, voting id, vote}

\item{weights}{if supplied it should be a named vector that converts votes into a numeric values that correspond to their similarity}

\item{allowMissings}{maximum number of missing votigs allowd for deputy.}
}
\value{
distance matrix
}
\description{
Function \code{get_distance_matrix} converts a data frame with tree columns (deputies' ids, votings' ids and votes) 
into a deputies distance matrix.
}
\details{
Function \code{get_distance_matrix} calculated distances among deputies based on their 
votes The more similar are the voting the smaller distance between deputies.
}
\examples{
# votes <- get_filtered_votes(terms_of_office = c(7,7))
data(votes)
v <- c(`Za` = 5, `Przeciw` = -5, `Wstrzymał się` = 2, `Nieobecny` = 0)/10
mat2 <- get_distance_matrix(votes[,c("surname_name", "id_voting", "vote")], weights = v)
\dontrun{
mat1 <- get_distance_matrix(votes[,c("surname_name", "id_voting", "vote")])
}

}
\author{
Przemyslaw Biecek
}

