% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_most_frequent_club.R
\name{get_most_frequent_club}
\alias{get_most_frequent_club}
\title{Gets the Most Frequent Club for Each Deputy}
\usage{
get_most_frequent_club(deputy_id, club)
}
\arguments{
\item{deputy_id}{a vector with deputy unique ids (may be also a vactor with characters name/surname)}

\item{club}{vector of length equal to \code{deputy_id} with club memberships for particular deputies}
}
\value{
a data frame with club memberships of deputies
}
\description{
One deputy may belong to many different clubs and change clubs over time. 
The function \code{get_most_frequent_club} calculates the most frequent club for each deputy.
}
\examples{
# votes <- get_filtered_votes(terms_of_office = c(7,7))
data(votes)
clubs <- get_most_frequent_club(votes$surname_name, votes$club)
head(clubs)

}
\author{
Przemyslaw Biecek
}
