% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_statements_table.R
\name{get_statements_table}
\alias{get_statements_table}
\title{Importing statements table from a database}
\usage{
get_statements_table(dbname = 'sejmrp', user = 'reader', password = 'qux94874',
host = 'services.mini.pw.edu.pl', sorted_by_id = TRUE)
}
\arguments{
\item{dbname}{name of database; default: 'sejmrp'}

\item{user}{name of user; default: 'reader'}

\item{password}{password of database; default: 'qux94874'}

\item{host}{name of host; default: 'services.mini.pw.edu.pl'}

\item{sorted_by_id}{information if table should be sorted by id; default: TRUE}
}
\value{
data frame
}
\description{
Function \code{get_statements_table} imports statements table from a database.
}
\note{
Default parameters use privilages of 'reader'. It can only SELECT data from database.

All information is stored in PostgreSQL database.
}
\examples{
\dontrun{
statements <- get_statements_table()
dim(statements)
# [1] 12629     3
names(statements)
# [1] "id_statement" "surname_name" "statement"}
}
\author{
Piotr Smuda
}

