% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_votes_table.R
\name{get_votes_table}
\alias{get_votes_table}
\title{Importing votes table from a database}
\usage{
get_votes_table(dbname = 'sejmrp', user = 'reader',
  password = 'qux94874', host = 'services.mini.pw.edu.pl',
  sorted_by_id = TRUE, windows = .Platform$OS.type == 'windows')
}
\arguments{
\item{dbname}{name of database; default: 'sejmrp'}

\item{user}{name of user; default: 'reader'}

\item{password}{password of database; default: 'qux94874'}

\item{host}{name of host; default: 'services.mini.pw.edu.pl'}

\item{sorted_by_id}{information if table should be sorted by id; default: TRUE}

\item{windows}{information of used operation system;
default: .Platform$OS.type == 'windows'}
}
\value{
data frame
}
\description{
Function \code{get_votes_table} imports votes table from a database.
}
\details{
Function \code{get_votes_table} imports votes table from a database.
The result of this function is a data frame with votes' data. Because of
encoding issue on Windows operation system, you need to select if you use Windows.
}
\note{
Default parameters use privilages of 'reader'. It can only SELECT data from database.

All information is stored in PostgreSQL database.
}
\examples{
\dontrun{
votes <- get_votes_table()
dim(votes)
# [1] 2741899       5
names(votes)
# [1] 'id_vote'   'id_deputy' 'id_voting' 'vote'      'club'
#' object.size(votes)
# 87767112 bytes}
}
\author{
Piotr Smuda
}

