% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/remove_database.R
\name{remove_database}
\alias{remove_database}
\title{Removing database}
\usage{
remove_database(dbname,user,password,host)
}
\arguments{
\item{dbname}{name of database}

\item{user}{name of user}

\item{password}{password of database}

\item{host}{name of host}
}
\value{
invisible NULL
}
\description{
Function \code{remove_database} remove whole database
}
\details{
// to do
}
\examples{
// to do
}
\author{
Piotr Smuda
}

