% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/srp_create_base.R
\name{srp_create_database}
\alias{srp_create_database}
\title{Creating database}
\usage{
srp_create_database(dbname,user,password,host)
}
\arguments{
\item{dbname}{name of database}

\item{user}{name of user}

\item{password}{password of database}

\item{host}{name of host}
}
\value{
invisible NULL
}
\description{
Function \code{srp_create_database} creates database with three empty
tables: deputies, votings, votes.
}
\details{
// to do
}
\examples{
// to do
}
\author{
Piotr Smuda
}

