% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/srp_deputies_add_new.R
\name{srp_deputies_add_new}
\alias{srp_deputies_add_new}
\title{Adding new deputies to table}
\usage{
srp_deputies_add_new(dbname,user,password,host,type,id)
}
\arguments{
\item{dbname}{name of database}

\item{user}{name of user}

\item{password}{password of database}

\item{host}{name of host}

\item{type}{type of deputies which be add to table with deputies: active, inactive}

\item{id}{id of deputies from which we start add new deputies}
}
\value{
invisible NULL
}
\description{
Function \code{srp_deputies_add_new} adds new deputies to table with deputies.
}
\details{
// to do
}
\examples{
// to do
}
\author{
Piotr Smuda
}

