% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/srp_deputies_data.R
\name{srp_deputies_data}
\alias{srp_deputies_data}
\title{Getting data about deputies}
\usage{
srp_deputies_data(type)
}
\arguments{
\item{type}{type of deputies which be add to table with deputies: active, inactive}
}
\value{
data frame with two columns: id_deputy, surname_name
}
\description{
Function \code{srp_deputies_data} gets data about deputies.
}
\details{
// to do
}
\examples{
// to do
}
\author{
Piotr Smuda
}

