% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/srp_deputies_table_update.R
\name{srp_deputies_table_update}
\alias{srp_deputies_table_update}
\title{Updating table with deputies}
\usage{
srp_deputies_table_update(dbname,user,password,host)
}
\arguments{
\item{dbname}{name of database}

\item{user}{name of user}

\item{password}{password of database}

\item{host}{name of host}
}
\value{
invisible NULL
}
\description{
Function \code{srp_deputies_table_update} updates table with deputies.
}
\details{
// to do
}
\examples{
// to do
}
\author{
Piotr Smuda
}

