% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statements_get_statement.R
\name{statements_get_statement}
\alias{statements_get_statement}
\title{Getting statements}
\usage{
statements_get_statement(page, ...)
}
\arguments{
\item{page}{deputy's statement's page}

\item{...}{other arguments, that will be passed to safe_html()}
}
\value{
character vector
}
\description{
Function \code{statements_get_statement} gets statement's content.
}
\details{
Function \code{statements_get_statement} gets statement's content.
Example of page with deputy's statement: 
http://www.sejm.gov.pl/Sejm7.nsf/wypowiedz.xsp?posiedzenie=15&dzien=1&wyp=008
}
\note{
All information is stored in PostgreSQL database.
}
\examples{
\dontrun{
page <- paste0('http://www.sejm.gov.pl/Sejm7.nsf/',
               'wypowiedz.xsp?posiedzenie=15&dzien=1&wyp=008')
statements_get_statement(page)}

}
\author{
Piotr Smuda, Tomasz Mikolajczyk
}
