% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statements_get_statements_table.R
\name{statements_get_statements_table}
\alias{statements_get_statements_table}
\title{Getting statements' table}
\usage{
statements_get_statements_table(page)
}
\arguments{
\item{page}{meeting's page}
}
\value{
data frame with three unnamed columns
}
\description{
Function \code{statements_get_statements_table} gets statements' table from
meeting's page.
}
\details{
Function \code{statements_get_statements_table} gets statements' table. from
meeting's page. Example of a meeting's page: 
http://www.sejm.gov.pl/Sejm7.nsf/posiedzenie.xsp?posiedzenie=99&dzien=2
The result of this function is a data frame with three columns, where
the first includes author of statement, the second the number of order point 
and the third is a title of order point.
}
\note{
All information is stored in PostgreSQL database.
}
\examples{
\dontrun{
page <- 'http://www.sejm.gov.pl/Sejm7.nsf/posiedzenie.xsp?posiedzenie=99&dzien=2'
statements_get_statements_table(page)}

}
\author{
Piotr Smuda
}
