% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votes.R
\docType{data}
\name{votes}
\alias{votes}
\title{Votes from 7th Office of Polish Sejm}
\format{2890479 rows and 9 columns}
\usage{
data(votes)
}
\description{
Votes taken in the 7th office of Polish Sejm (2011-2015)
\itemize{
   \item surname_name Surname and name of a deputy
   \item nr_term_of_office Which office? In this sample there is only 7th office
   \item club club of the deputy at the moment of voting (may change in time)
   \item vote vote taken in the voting
   \item id_voting unique id of the voting
   \item nr_meeting no of the meeting
   \item nr_voting no of the voting
   \item date_meeting data of the meeting
   \item topic_voting full title of the voting
}
}
\keyword{datasets}

