% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/votes_create_table.R
\name{votes_create_table}
\alias{votes_create_table}
\title{Creating table with votes}
\usage{
votes_create_table(dbname,user,password,host,home_page,windows=TRUE)
}
\arguments{
\item{dbname}{name of database}

\item{user}{name of user}

\item{password}{password of database}

\item{host}{name of host}

\item{home_page}{main page of polish diet: http://www.sejm.gov.pl/Sejm7.nsf/}

\item{windows}{information of used operation system; default: TRUE}
}
\value{
invisible NULL
}
\description{
Function \code{votes_create_table} creates table with votes.
}
\details{
// to do
}
\examples{
// to do
}
\author{
Piotr Smuda
}

