% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/votes_get_clubs_links.R
\name{votes_get_clubs_links}
\alias{votes_get_clubs_links}
\title{Getting links with voting's results for each club}
\usage{
votes_get_clubs_links(home_page,page)
}
\arguments{
\item{home_page}{main page of polish diet: http://www.sejm.gov.pl/Sejm7.nsf/}

\item{page}{voting's page}
}
\value{
data frame with two columns: club, links
}
\description{
Function \code{votes_get_clubs_links} gets links with voting's results for each club.
}
\details{
// to do
}
\examples{
// to do
}
\author{
Piotr Smuda
}

