% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/votes_get_results.R
\name{votes_get_results}
\alias{votes_get_results}
\title{Getting voting's results for each club}
\usage{
votes_get_results(page)
}
\arguments{
\item{page}{club's voting's results page}
}
\value{
data frame with two columns: deputy, vote
}
\description{
Function \code{votes_get_results} gets voting's results for each club.
}
\details{
// to do
}
\examples{
// to do
}
\author{
Piotr Smuda
}

