% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/votings_get_date.R
\name{votings_get_date}
\alias{votings_get_date}
\title{Getting date of meeting}
\usage{
votings_get_date(page)
}
\arguments{
\item{page}{meeting's page}
}
\value{
date in format YYYY-MM-DD as character
}
\description{
Function \code{votings_get_date} gets a date of meeting.
}
\details{
Example of a meeting's page:
http://www.sejm.gov.pl/Sejm7.nsf/agent.xsp?symbol=listaglos&IdDnia=1179
}
\note{
All information is stored in PostgreSQL database.
}
\examples{
\dontrun{
page <- 'http://www.sejm.gov.pl/Sejm7.nsf/agent.xsp?symbol=listaglos&IdDnia=1179'
votings_get_date(page)}
}
\author{
Piotr Smuda
}

