% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/votings_get_meetings_links.R
\name{votings_get_meetings_links}
\alias{votings_get_meetings_links}
\title{Getting meetings' links}
\usage{
votings_get_meetings_links(home_page,page)
}
\arguments{
\item{home_page}{main page of polish diet: http://www.sejm.gov.pl/Sejm7.nsf/}

\item{page}{page with votings in polish diet: http://www.sejm.gov.pl/Sejm7.nsf/agent.xsp?symbol=posglos&NrKadencji=7}
}
\value{
character vector
}
\description{
Function \code{votings_get_meetings_links} gets meetings' links.
}
\details{
// to do
}
\examples{
// to do
}
\author{
Piotr Smuda
}

