% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/votings_get_meetings_table.R
\name{votings_get_meetings_table}
\alias{votings_get_meetings_table}
\title{Getting meetings' table}
\usage{
votings_get_meetings_table(page =
  'http://www.sejm.gov.pl/Sejm7.nsf/agent.xsp?symbol=posglos&NrKadencji=7')
}
\arguments{
\item{page}{page with votings in polish diet:
http://www.sejm.gov.pl/Sejm7.nsf/agent.xsp?
symbol=posglos&NrKadencji=7}
}
\value{
data frame with three unnamed columns
}
\description{
Function \code{votings_get_meetings_table} gets meetings' table.
}
\details{
Function \code{votings_get_meetings_table} gets meetings' table. The
result of this function is a data frame with three columns, where
the first includes numbers of meetings, the second theirs' dates in
Polish and the third is with numbers of votings on each meeting.
}
\note{
All information is stored in PostgreSQL database.
}
\examples{
\dontrun{
votings_get_meetings_table()}
}
\author{
Piotr Smuda
}

