% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/votings_get_votings_links.R
\name{votings_get_votings_links}
\alias{votings_get_votings_links}
\title{Getting votings' links}
\usage{
votings_get_votings_links(home_page,page)
}
\arguments{
\item{home_page}{main page of polish diet: http://www.sejm.gov.pl/Sejm7.nsf/}

\item{page}{meeting's page}
}
\value{
character vector
}
\description{
Function \code{votings_get_votings_links} gets votings' links.
}
\details{
// to do
}
\examples{
// to do
}
\author{
Piotr Smuda
}

