% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/votings_get_votings_table.R
\name{votings_get_votings_table}
\alias{votings_get_votings_table}
\title{Getting votings' table}
\usage{
votings_get_votings_table(page)
}
\arguments{
\item{page}{meeting's page}
}
\value{
data frame with three columns: Nr, Godzina (Time), Temat (Topic)
}
\description{
Function \code{votings_get_votings_table} gets votings' table.
}
\details{
// to do
}
\examples{
// to do
}
\author{
Piotr Smuda
}

