% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/votings_update_table.R
\name{votings_update_table}
\alias{votings_update_table}
\title{Updating table with votings}
\usage{
votings_update_table(dbname, user, password, host,
  home_page='http://www.sejm.gov.pl/Sejm7.nsf/', page=
  'http://www.sejm.gov.pl/Sejm7.nsf/agent.xsp?symbol=posglos&NrKadencji=7',
  verbose=FALSE)
}
\arguments{
\item{dbname}{name of database}

\item{user}{name of user}

\item{password}{password of database}

\item{host}{name of host}

\item{home_page}{main page of polish diet: http://www.sejm.gov.pl/Sejm7.nsf/}

\item{page}{page with votings in polish diet:
http://www.sejm.gov.pl/Sejm7.nsf/agent.xsp?
symbol=posglos&NrKadencji=7}

\item{verbose}{if TRUE then additional info will be printed}
}
\value{
invisible NULL
}
\description{
Function \code{votings_update_table} updates table with votings.
}
\note{
All information is stored in PostgreSQL database.
}
\examples{
\dontrun{
votings_update_table(dbname, user, password, host)}
}
\author{
Piotr Smuda
}

