% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deputies_create_table.R
\name{deputies_create_table}
\alias{deputies_create_table}
\title{Creating table with deputies}
\usage{
deputies_create_table(dbname, user, password, host)
}
\arguments{
\item{dbname}{name of database}

\item{user}{name of user}

\item{password}{password of database}

\item{host}{name of host}
}
\value{
invisible NULL
}
\description{
Function \code{deputies_create_table} creates a table with deputies.
}
\note{
Use only this function for first time, when the \emph{deputies} table
is empty. Then use \code{deputies_update_table}.

All information is stored in PostgreSQL database.
}
\examples{
\dontrun{
deputies_create_table(dbname, user, password, host)}
}
\author{
Piotr Smuda
}

