% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deputies_get_ids.R
\name{deputies_get_ids}
\alias{deputies_get_ids}
\title{Getting deputies' ids}
\usage{
deputies_get_ids(dbname,user,password,host,windows)
}
\arguments{
\item{dbname}{name of database}

\item{user}{name of user}

\item{password}{password of database}

\item{host}{name of host}

\item{windows}{information of used operation system; default: TRUE}
}
\value{
named character vector
}
\description{
Function \code{deputies_get_ids} gets deputies' ids.
}
\details{
// to do
}
\examples{
// to do
}
\author{
Piotr Smuda
}

