% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deputies_update_table.R
\name{deputies_update_table}
\alias{deputies_update_table}
\title{Updating table with deputies}
\usage{
deputies_update_table(dbname, user, password, host)
}
\arguments{
\item{dbname}{name of database}

\item{user}{name of user}

\item{password}{password of database}

\item{host}{name of host}
}
\value{
invisible NULL
}
\description{
Function \code{deputies_update_table} updates a table with deputies.
}
\note{
All information is stored in PostgreSQL database.
}
\examples{
\dontrun{
deputies_update_table(dbname, user, password, host)}
}
\author{
Piotr Smuda
}

