% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deputies_update_table.R
\name{deputies_update_table}
\alias{deputies_update_table}
\title{Updating table with deputies}
\usage{
deputies_update_table(dbname, user, password, host,
   nr_term_of_office = 8)
}
\arguments{
\item{dbname}{name of database}

\item{user}{name of user}

\item{password}{password of database}

\item{host}{name of host}

\item{nr_term_of_office}{number of term of office of Polish Diet; default: 8}
}
\value{
invisible NULL
}
\description{
Function \code{deputies_update_table} updates a table with deputies.
}
\note{
All information is stored in PostgreSQL database.
}
\examples{
\dontrun{
deputies_update_table(dbname, user, password, host)}

}
\author{
Piotr Smuda
}
