% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_database.R
\name{remove_database}
\alias{remove_database}
\title{Removing database}
\usage{
remove_database(dbname, user, password, host)
}
\arguments{
\item{dbname}{name of database}

\item{user}{name of user}

\item{password}{password of database}

\item{host}{name of host}
}
\value{
invisible NULL
}
\description{
Function \code{remove_database} remove whole database.
}
\note{
All information is stored in PostgreSQL database.
}
\examples{
\dontrun{
remove_database(dbname, user, password, host)}

}
\author{
Piotr Smuda
}
