% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/statements_create_table.R
\name{statements_create_table}
\alias{statements_create_table}
\title{Creating table with deputies' statements}
\usage{
statements_create_table(dbname, user, password, host,
  home_page = 'http://www.sejm.gov.pl/Sejm7.nsf/')
}
\arguments{
\item{dbname}{name of database}

\item{user}{name of user}

\item{password}{password of database}

\item{host}{name of host}

\item{home_page}{main page of polish diet: http://www.sejm.gov.pl/Sejm7.nsf/}
}
\value{
invisible NULL
}
\description{
Function \code{statements_create_table} creates a table with deputies' statements.
}
\note{
Use only this function for first time, when the \emph{statements} table
is empty. Then use \code{statements_update_table}.

All information is stored in PostgreSQL database.
}
\examples{
\dontrun{
statements_create_table(dbname, user, password, host)}
}
\author{
Piotr Smuda, Tomasz Mikolajczyk
}

