% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/statements_update_table.R
\name{statements_update_table}
\alias{statements_update_table}
\title{Updating table with deputies' statements}
\usage{
statements_update_table(dbname,user,password,host,home_page="http://www.sejm.gov.pl/Sejm7.nsf/")
}
\arguments{
\item{dbname}{name of database}

\item{user}{name of user}

\item{password}{password of database}

\item{host}{name of host}

\item{home_page}{main page of polish diet: http://www.sejm.gov.pl/Sejm7.nsf/}
}
\value{
invisible NULL
}
\description{
Function \code{statements_update_table} updates a table with deputies' statements.
}
\note{
All information is stored in PostgreSQL database.
}
\examples{
\dontrun{
statements_update_table(dbname,user,password,host)}
}
\author{
Piotr Smuda, Tomasz Mikolajczyk
}

