% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/votes_match_deputies_ids.R
\name{votes_match_deputies_ids}
\alias{votes_match_deputies_ids}
\title{Matching deputies to theirs' ids}
\usage{
votes_match_deputies_ids(dbname,user,password,host,page,windows=TRUE)
}
\arguments{
\item{dbname}{name of database}

\item{user}{name of user}

\item{password}{password of database}

\item{host}{name of host}

\item{page}{club's voting's results page}

\item{windows}{information of used operation system; default: TRUE}
}
\value{
invisible NULL
}
\description{
Function \code{votes_match_deputies_ids} matches deputies to theirs' ids.
}
\details{
// to do
}
\examples{
// to do
}
\author{
Piotr Smuda
}

