% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/votings_create_table.R
\name{votings_create_table}
\alias{votings_create_table}
\title{Creating table with votings}
\usage{
votings_create_table(dbname,user,password,host,
  home_page="http://www.sejm.gov.pl/Sejm7.nsf/",page=
  "http://www.sejm.gov.pl/Sejm7.nsf/agent.xsp?symbol=posglos&NrKadencji=7")
}
\arguments{
\item{dbname}{name of database}

\item{user}{name of user}

\item{password}{password of database}

\item{host}{name of host}

\item{home_page}{main page of polish diet: http://www.sejm.gov.pl/Sejm7.nsf/}

\item{page}{page with votings in polish diet:
http://www.sejm.gov.pl/Sejm7.nsf/agent.xsp?
symbol=posglos&NrKadencji=7}
}
\value{
invisible NULL
}
\description{
Function \code{votings_create_table} creates a table with votings.
}
\note{
Use only this function for first time, when the \emph{votings} table
is empty. Then use \code{votings_update_table}.

All information is stored in PostgreSQL database.
}
\examples{
\dontrun{
votings_create_table(dbname,user,password,host)}
}
\author{
Piotr Smuda
}

