% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/votings_get_date.R
\name{votings_get_date}
\alias{votings_get_date}
\title{Getting date of meeting}
\usage{
votings_get_date(page)
}
\arguments{
\item{page}{meeting's page}
}
\value{
date in format YYYY-MM-DD as character
}
\description{
Function \code{votings_get_date} gets date of meeting.
}
\details{
// to do
}
\examples{
// to do
}
\author{
Piotr Smuda
}

