% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votings_get_meetings_links.R
\name{votings_get_meetings_links}
\alias{votings_get_meetings_links}
\title{Getting meetings' links}
\usage{
votings_get_meetings_links(
  home_page = 'http://www.sejm.gov.pl/Sejm8.nsf/', page = 
  'http://www.sejm.gov.pl/Sejm8.nsf/agent.xsp?symbol=posglos&NrKadencji=8')
}
\arguments{
\item{home_page}{main page of polish diet: http://www.sejm.gov.pl/Sejm8.nsf/}

\item{page}{page with votings in polish diet: 
http://www.sejm.gov.pl/Sejm8.nsf/agent.xsp?
symbol=posglos&NrKadencji=8}
}
\value{
character vector
}
\description{
Function \code{votings_get_meetings_links} gets meetings' links.
}
\note{
All information is stored in PostgreSQL database.
}
\examples{
\dontrun{
votings_get_meetings_links()}

}
\author{
Piotr Smuda
}

