% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/votings_get_meetings_table.R
\name{votings_get_meetings_table}
\alias{votings_get_meetings_table}
\title{Getting meetings' table}
\usage{
votings_get_meetings_table(page)
}
\arguments{
\item{page}{page with votings in polish diet: http://www.sejm.gov.pl/Sejm7.nsf/agent.xsp?symbol=posglos&NrKadencji=7}
}
\value{
data frame with three unnamed columns
}
\description{
Function \code{votings_get_meetings_table} gets meetings' table.
}
\details{
// to do
}
\examples{
// to do
}
\author{
Piotr Smuda
}

