% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/votings_get_votings_table.R
\name{votings_get_votings_table}
\alias{votings_get_votings_table}
\title{Getting votings' table}
\usage{
votings_get_votings_table(page)
}
\arguments{
\item{page}{meeting's page}
}
\value{
data frame with three columns: Nr, Godzina (Time), Temat (Topic)
}
\description{
Function \code{votings_get_votings_table} gets votings' table from
meeting's page.
}
\details{
Function \code{votings_get_votings_table} gets votings' table from
meeting's page. Example of a meeting's page:
http://www.sejm.gov.pl/Sejm7.nsf/agent.xsp?symbol=listaglos&IdDnia=1179
The result of this function is a data frame with three columns, where
the first includes numbers of votings, the second voting's time
and the third is with voting's topics.
}
\note{
All information is stored in PostgreSQL database.
}
\examples{
\dontrun{
page <- 'http://www.sejm.gov.pl/Sejm7.nsf/agent.xsp?symbol=listaglos&IdDnia=1179'
votings_get_votings_table(page)}
}
\author{
Piotr Smuda
}

